Rename coordinators and update imports
